Initial Commit of demo with CODEC2
CODEC2 added to symposium demo.
Frequency change from 2m to 70cm.
Changed payload_size from 100 to 64 to match the native block size of
CODEC2.
Added Vector-to-Stream after the CODEC2 encoder, and Stream-to-Vector
before the CODEC2 decoder, to match CODEC2 up to the byte stream
handling.
Changed hard-coded multiples of 110 in the channel selector widget to
expressions in terms of payload_size and guard_size, so they now
auto-adjust to those variations.
Passes a brief test.